folder_branch_ops: cache flushed MDs
When a journal converts its branch, those converted MDs should go in
the local MDCache -- otherwise, they will disappear when the journal
is flushed, and CR will have to fetch them from the server.

So this commit passes the flushed BareRootMetadatas to folderBranchOps
via mdFlushListener, which is equipped to convert them to something
that can be cached.

Note that I didn't factor out similar code that already exists for
this in MDOpsStandard and journalMDOps, because that code does more
checks than we need, and expects some additional info we don't have
here.

Issue: KBFS-1574